Crate acvm_blackbox_solver

This crate provides the implementation of BlackBox functions of ACIR and Brillig. For functions that are backend-dependent, it provides a Trait BlackBoxFunctionSolver that must be implemented by the backend. For functions that have a reference implementation, such as keccakf1600, this crate exports the reference implementation directly.

Modules

aes128 🔒

curve_specific_solver 🔒

ecdsa 🔒

hash 🔒

logic 🔒

Structs

StubbedBlackBoxSolver

Enums

BlackBoxResolutionError

Traits

BlackBoxFunctionSolver

This component will generate outputs for Blackbox function calls where the underlying [acir::BlackBoxFunc] doesn’t have a canonical Rust implementation.

Functions

aes128_encrypt

bit_and

bit_xor

blake3

blake2s

ecdsa_secp256k1_verify

ecdsa_secp256r1_verify

keccakf1600

Keccak permutation for a state of size 1600 bits, represented by 25 lanes of 64 bits (25*64 = 1600)

sha256_compression